fix id condition not used bug
